+++ modules/student.js
@@ -35,33 +35,6 @@ class Student {
 	}
 }
 
-async function passwordRequest(newPassword, username) {
-	if (newPassword && username) {
-		let passwordChange = true;
-		io.emit("passwordUpdate", passwordChange, username, newPassword);
-	};
-};
-
-function getUserClass(username) {
-	try {
-		logger.log('info', `[getUserClass] username=(${username})`)
-
-		for (let classCode of Object.keys(classInformation)) {
-			if (classInformation[classCode].students[username]) {
-				logger.log('verbose', `[getUserClass] classCode=(${classCode})`)
-				return classCode
-			}
-		}
-
-		logger.log('verbose', `[getUserClass] classCode=(${null})`)
-		return null
-	} catch (err) {
-		return err
-	}
-}
-
 module.exports = {
-	Student,
-    passwordRequest,
-	getUserClass
+	Student
 }
